src/jest/Cpu.test.ts
import { describe, expect, test } from '@jest/globals';
import { CPU, Memory, OPCODES } from '../cpu/Cpu';

let cpu = new CPU();
let mem = new Memory();
describe('CPU LDA tests', () => {
  test('LDA_IMMEDIATE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDA_IMMEDIATE;
    mem.memory[0x1] = 0x30;
    cpu.cycles = 2;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x30);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ZEROPAGE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDA_ZEROPAGE;
    mem.memory[0x1] = 0x42;
    mem.memory[0x0042] = 0x84;
    cpu.cycles = 3;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x84);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ZEROPAGE_X, noWrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f; //hardcode x
    mem.memory[0x0] = OPCODES.LDA_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    mem.memory[0x0042 + 0x0f] = 0x55;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ZEROPAGE_X, wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f; //hardcode x
    mem.memory[0x0] = OPCODES.LDA_ZEROPAGE_X;
    mem.memory[0x1] = 0xff;
    mem.memory[0xe] = 0x55; // 0xff + 0x0f + wrap = 0xe
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ABSOLUTE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDA_ABSOLUTE;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5542] = 0x69;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ABSOLUTE_X, no Page cross', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f; //hardcode x
    mem.memory[0x0] = OPCODES.LDA_ABSOLUTE_X;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5551] = 0x69; //0x5542 + 0x0f
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ABSOLUTE_X,Page cross', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x1; //hardcode x
    mem.memory[0x0] = OPCODES.LDA_ABSOLUTE_X;
    mem.memory[0x1] = 0xff;
    mem.memory[0x2] = 0x44; //0x4402
    mem.memory[0x4500] = 0x69; //0x4402 + 0x1, page cross
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ABSOLUTE_Y, no Page cross', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x0f; //hardcode Y
    mem.memory[0x0] = OPCODES.LDA_ABSOLUTE_Y;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5551] = 0x69; //0x5542 + 0x0f
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_ABSOLUTE_Y,Page cross', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x1; //hardcode Y
    mem.memory[0x0] = OPCODES.LDA_ABSOLUTE_Y;
    mem.memory[0x1] = 0xff;
    mem.memory[0x2] = 0x44; //0x44ff
    mem.memory[0x4500] = 0x69; //0x44ff + 0x1, page cross
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_INDEXED_INDIRECT_X, noWrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x10; //hardcode X
    mem.memory[0x0] = OPCODES.LDA_INDEXED_INDIRECT_X;
    mem.memory[0x1] = 0x4;
    mem.memory[0x14] = 0x69; // no wrap
    mem.memory[0x15] = 0x32; // 0x3269
    mem.memory[0x3269] = 0x44;
    cpu.cycles = 6;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x44);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_INDEXED_INDIRECT_X, wrap around', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x7; //hardcode X
    mem.memory[0x0] = OPCODES.LDA_INDEXED_INDIRECT_X;
    mem.memory[0x1] = 0xff;
    mem.memory[0x6] = 0x69; //wrap
    mem.memory[0x7] = 0x44; //0x4469
    mem.memory[0x4469] = 0x55;
    cpu.cycles = 6;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_INDIRECT_INDEXED_Y, page cross', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0xff; //hardcode Y
    mem.memory[0x0] = OPCODES.LDA_INDIRECT_INDEXED_Y;
    mem.memory[0x1] = 0x10;
    mem.memory[0x10] = 0x01; //wrap
    mem.memory[0x11] = 0x80; //wrap
    mem.memory[0x8100] = 0x55;
    cpu.cycles = 6;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDA_INDIRECT_INDEXED_Y, no page cross', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x0; //hardcode Y
    mem.memory[0x0] = OPCODES.LDA_INDIRECT_INDEXED_Y;
    mem.memory[0x1] = 0x10;
    mem.memory[0x10] = 0x00; //wrap
    mem.memory[0x11] = 0x02; //wrap
    mem.memory[0x200] = 0x55;
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
});
describe('CPU LDX tests', () => {
  test('LDX_IMMEDIATE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDX_IMMEDIATE;
    mem.memory[0x1] = 0x30;
    cpu.cycles = 2;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x30);
    expect(cpu.cycles).toBe(0);
  });
  test('LDX_ZEROPAGE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDX_ZEROPAGE;
    mem.memory[0x1] = 0x42;
    mem.memory[0x0042] = 0x84;
    cpu.cycles = 3;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x84);
    expect(cpu.cycles).toBe(0);
  });
  test('LDX_ZEROPAGE_Y,wrap', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0xff;
    mem.memory[0x0] = OPCODES.LDX_ZEROPAGE_Y;
    mem.memory[0x1] = 0x42;
    mem.memory[0x41] = 0x55; //0x42 + 0xff + wrap = 0x41
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDX_ZEROPAGE_Y, no wrap', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x0f;
    mem.memory[0x0] = OPCODES.LDX_ZEROPAGE_Y;
    mem.memory[0x1] = 0x42;
    mem.memory[0x0042 + 0x0f] = 0x55;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDX_ABSOLUTE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDX_ABSOLUTE;
    mem.memory[0x1] = 0x33;
    mem.memory[0x2] = 0x44; //0x4433
    mem.memory[0x4433] = 0x69;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDX_ABSOLUTE_Y,Page cross', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x1; //hardcode x
    mem.memory[0x0] = OPCODES.LDX_ABSOLUTE_Y;
    mem.memory[0x1] = 0xff;
    mem.memory[0x2] = 0x44; //0x4402
    mem.memory[0x4500] = 0x69; //0x4402 + 0x1, page cross
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDX_ABSOLUTE_Y, no Page cross', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x0f; //hardcode Y
    mem.memory[0x0] = OPCODES.LDX_ABSOLUTE_Y;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5551] = 0x69; //0x5542 + 0x0f
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.X).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
});
describe('CPU STA tests', () => {
  test('STA_ZEROPAGE', () => {
    cpu.reset(mem);
    cpu.registers.A = 0x13;
    mem.memory[0x0] = OPCODES.STA_ZEROPAGE;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 3;
    cpu.execute(mem);
    expect(mem.memory[0x42]).toEqual(0x13);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_ZEROPAGE_X,wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0xff;
    cpu.registers.A = 0x13;
    mem.memory[0x0] = OPCODES.STA_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x41]).toEqual(0x13);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_ZEROPAGE_X, no wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f;
    cpu.registers.A = 0x13;
    mem.memory[0x0] = OPCODES.STA_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x042 + 0x0f]).toEqual(0x13);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_ABSOLUTE', () => {
    cpu.reset(mem);
    cpu.registers.A = 0x11;
    mem.memory[0x0] = OPCODES.STA_ABSOLUTE;
    mem.memory[0x1] = 0x33;
    mem.memory[0x2] = 0x44; //0x4433
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x4433]).toEqual(0x11);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_ABSOLUTE_X', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f;
    cpu.registers.A = 0x35;
    mem.memory[0x0] = OPCODES.STA_ABSOLUTE_X;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5551] = 0x69; //0x5542 + 0x0f
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(mem.memory[0x5551]).toEqual(0x35);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_ABSOLUTE_Y', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x0f;
    cpu.registers.A = 0x11;
    mem.memory[0x0] = OPCODES.STA_ABSOLUTE_Y;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5551] = 0x69; //0x5542 + 0x0f
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(mem.memory[0x5551]).toEqual(0x11);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_INDEXED_INDIRECT_X, wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0xff;
    cpu.registers.A = 0x69;
    mem.memory[0x0] = OPCODES.STA_INDEXED_INDIRECT_X;
    mem.memory[0x1] = 0xff;
    mem.memory[0xfe] = 0x69; //wrap
    mem.memory[0xff] = 0x44; //0x4469
    mem.memory[0x4469] = 0x55;
    cpu.cycles = 6;
    cpu.execute(mem);
    expect(mem.memory[0xfe]).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_INDEXED_INDIRECT_X, nowrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x1;
    cpu.registers.A = 0x69;
    mem.memory[0x0] = OPCODES.STA_INDEXED_INDIRECT_X;
    mem.memory[0x1] = 0x05;
    mem.memory[0x6] = 0x69;
    mem.memory[0x7] = 0x44; //0x4469
    cpu.cycles = 6;
    cpu.execute(mem);
    expect(mem.memory[0x4469]).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('STA_INDIRECT_INDEXED_Y', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0xff;
    cpu.registers.A = 0x69;
    mem.memory[0x0] = OPCODES.STA_INDIRECT_INDEXED_Y;
    mem.memory[0x1] = 0x10;
    mem.memory[0x10] = 0x01;
    mem.memory[0x11] = 0x80; //0x8001 + A
    cpu.cycles = 6;
    cpu.execute(mem);
    expect(mem.memory[0x8100]).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
});
describe('CPU STX tests', () => {
  test('STX_ZEROPAGE', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x55;
    mem.memory[0x0] = OPCODES.STX_ZEROPAGE;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 3;
    cpu.execute(mem);
    expect(mem.memory[0x42]).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('STX_ZEROPAGE_Y,wrap', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0xff;
    cpu.registers.X = 0x43;
    mem.memory[0x0] = OPCODES.STX_ZEROPAGE_Y;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x41]).toEqual(0x43);
    expect(cpu.cycles).toBe(0);
  });
  test('STX_ZEROPAGE_Y, no wrap', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x0f;
    cpu.registers.X = 0x73;
    mem.memory[0x0] = OPCODES.STX_ZEROPAGE_Y;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x042 + 0x0f]).toEqual(0x73);
    expect(cpu.cycles).toBe(0);
  });
  test('STX_ABSOLUTE', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x42;
    mem.memory[0x0] = OPCODES.STX_ABSOLUTE;
    mem.memory[0x1] = 0x33;
    mem.memory[0x2] = 0x44; //0x4433
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x4433]).toEqual(0x42);
    expect(cpu.cycles).toBe(0);
  });
});
describe('CPU STY tests', () => {
  test('STY_ZEROPAGE', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x44;
    mem.memory[0x0] = OPCODES.STY_ZEROPAGE;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 3;
    cpu.execute(mem);
    expect(mem.memory[0x42]).toEqual(0x44);
    expect(cpu.cycles).toBe(0);
  });
  test('STY_ZEROPAGE_X,wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0xff;
    cpu.registers.Y = 0x23;
    mem.memory[0x0] = OPCODES.STY_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x41]).toEqual(0x23);
    expect(cpu.cycles).toBe(0);
  });
  test('STY_ZEROPAGE_X, no wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f;
    cpu.registers.Y = 0x53;
    mem.memory[0x0] = OPCODES.STY_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x042 + 0x0f]).toEqual(0x53);
    expect(cpu.cycles).toBe(0);
  });
  test('STY_ABSOLUTE', () => {
    cpu.reset(mem);
    cpu.registers.Y = 0x69;
    mem.memory[0x0] = OPCODES.STY_ABSOLUTE;
    mem.memory[0x1] = 0x33;
    mem.memory[0x2] = 0x44; //0x4433
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(mem.memory[0x4433]).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
});
describe('CPU LD flags tests', () => {
  test('Zero flag set when byte is zero', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDA_IMMEDIATE;
    // mem.memory[0x1] = 0x30;
    cpu.cycles = 2;
    cpu.execute(mem);
    expect(cpu.cycles).toBe(0);
    expect(cpu.flags.Z).toBe(true);
  });
  test('Zero flag cleared when byte is not zero', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDA_IMMEDIATE;
    mem.memory[0x1] = 0x30;
    cpu.cycles = 2;
    cpu.execute(mem);
    expect(cpu.cycles).toBe(0);
    expect(cpu.flags.Z).toBe(false);
  });
});
describe('CPU LDY tests', () => {
  test('LDY_IMMEDIATE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDY_IMMEDIATE;
    mem.memory[0x1] = 0x30;
    cpu.cycles = 2;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x30);
    expect(cpu.cycles).toBe(0);
  });
  test('LDY_ZEROPAGE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDY_ZEROPAGE;
    mem.memory[0x1] = 0x42;
    mem.memory[0x0042] = 0x84;
    cpu.cycles = 3;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x84);
    expect(cpu.cycles).toBe(0);
  });
  test('LDY_ZEROPAGE_X,wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0xff;
    mem.memory[0x0] = OPCODES.LDY_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    mem.memory[0x41] = 0x55; //0x42 + 0xff + wrap = 0x41
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDY_ZEROPAGE_X, no wrap', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f;
    mem.memory[0x0] = OPCODES.LDY_ZEROPAGE_X;
    mem.memory[0x1] = 0x42;
    mem.memory[0x0042 + 0x0f] = 0x55;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x55);
    expect(cpu.cycles).toBe(0);
  });
  test('LDY_ABSOLUTE', () => {
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.LDY_ABSOLUTE;
    mem.memory[0x1] = 0x33;
    mem.memory[0x2] = 0x44; //0x4433
    mem.memory[0x4433] = 0x69;
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDY_ABSOLUTE_X,Page cross', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x1;
    mem.memory[0x0] = OPCODES.LDY_ABSOLUTE_X;
    mem.memory[0x1] = 0xff;
    mem.memory[0x2] = 0x44; //0x4402
    mem.memory[0x4500] = 0x69; //0x4402 + 0x1, page cross
    cpu.cycles = 5;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
  test('LDY_ABSOLUTE_X, no Page cross', () => {
    cpu.reset(mem);
    cpu.registers.X = 0x0f; //hardcode Y
    mem.memory[0x0] = OPCODES.LDY_ABSOLUTE_X;
    mem.memory[0x1] = 0x42;
    mem.memory[0x2] = 0x55; //0x5542
    mem.memory[0x5551] = 0x69; //0x5542 + 0x0f
    cpu.cycles = 4;
    cpu.execute(mem);
    expect(cpu.registers.Y).toEqual(0x69);
    expect(cpu.cycles).toBe(0);
  });
});

describe('CPU Jump tests', () => {
  test('JSR + LDA_IMMEDIATE', () => {
    // let cpu = new CPU();
    // let mem = new Memory();
    cpu.reset(mem);
    mem.memory[0x0] = OPCODES.JSR_ABSOLUTE;
    mem.memory[0x1] = 0x30;
    mem.memory[0x2] = 0x30;
    mem.memory[0x3030] = OPCODES.LDA_IMMEDIATE;
    mem.memory[0x3031] = 0x66;
    cpu.cycles = 8;
    cpu.execute(mem);
    expect(cpu.registers.A).toEqual(0x66); //test that jsr jumps to LDA and loads A
    expect(cpu.PC).toEqual(0x3032); //pc should be after LDA immediate
    expect(cpu.cycles).toBe(0); //cycles used should be 6 + 2
    expect(cpu.SP).toBe(0xff - 1); //stackpointer should be decremented (jos initialize vaihtunu ni failaa)
    let word = mem.memory[0xff] | (mem.memory[0xff + 1] << 8);
    expect(word).toEqual(0x2); //PC should be stored in memory at SP
  });
});
